+++ packages/client/src/base/source-uri-aware.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2020-2022 EclipseSource and others.
+ * Copyright (c) 2020-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,10 +14,17 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { AnyObject, hasStringProp } from '@eclipse-glsp/protocol';
+/**
+ * @deprecated No longer in use. Might be removed in future versions
+ */
 export interface SourceUriAware {
     sourceURI: string;
 }
 
+/**
+ * @deprecated No longer in use. Might be removed in future versions
+ */
+// eslint-disable-next-line deprecation/deprecation
 export function isSourceUriAware(obj: any): obj is SourceUriAware {
     return AnyObject.is(obj) && hasStringProp(obj, 'sourceURI');
 }